Split out exclusion list test after rebase
